Handle Android Multi Window guidelines
"In multi-window mode, an app can be in the paused state and still be visible to the user. An app might need to continue its activities even while paused. For example, a video-playing app that is in paused mode but is visible should continue showing its video. For this reason, we recommend that activities that play video not pause the video in their onPause() handlers. 

Instead, they should pause video in onStop(), and resume playback in onStart(). "